feature: break up files, allow for arbitrary-length arg arrays (#42)
* feature: break up files, allow for arbitrary-length arg arrays

* bug: outputs don't have outpoints
